stage0: use runRktAndGetUUID in TestPodManifest
Moved runRktAndGetUUID to tests/rkt_tests.go, and modified
TestPodManifest to use it. Reduces code duplication.
